changing class components to function components- having trouble with accessing endpoint
