+++ client/package.json
@@ -12,11 +12,14 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "axios": "^0.21.0",
+    "core-js": "^3.8.1",
     "dotenv": "^8.2.0",
     "prop-types": "^15.7.2",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
+    "regenerator-runtime": "^0.13.7",
     "semantic-ui-react": "^2.0.1",
     "styled-components": "^5.2.1"
   },
