+++ client/src/App/App.js
@@ -1,24 +1,17 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import Home from './pages/Home';
 import List from "./pages/List";
 
-class App extends Component {
-  render() {
-    const App = () => (
-      <div>
-        <Switch>
-          <Route exact path='/' component={Home}/>
-          <Route path='/list' component={List}/>
-        </Switch>
-      </div>
-    )
-    return(
+function App() {
+  return (
+    <div>
       <Switch>
-        <App/>
+        <Route exact path='/' component={Home}/>
+        <Route path='/list' component={List}/>
       </Switch>
-    )
-  }
+    </div> 
+  )
 }
 
 export default App;
