+++ client/src/App/pages/List.js
@@ -1,51 +1,47 @@
-import React, { Component } from "react";
+// List.js will only be used for testing purposes
+// for connecting to api/server.js
+import React, { useState, useEffect }from 'react';
 import { CenterContainer } from "../styles/centeredContainer";
-import {Title} from '../styles/title';
+import { Title } from '../styles/title';
+import axios from 'axios';
 
-class List extends Component {
-  // initialize the state
-  constructor(props){
-    super(props);
-    this.state = {
-      list: []
-    }
-  }
+function List() {
+  const [list, setList] = useState({listArr: []});
   
-  // Fetch the list 
-  componentDidMount() {
-    fetch('http://localhost:8000/api/getList')
-    .then((res) => res.json())
-    .then(list => this.setState({ list }));
-  }
-
-
-  render() {
-    const { list } = this.state;
-
-    return(
-      <div className="App">
-        <Title>List of Items</Title>
-        {/** Check to see if any items are found */}
-        {list.length ? (
-          <CenterContainer>
-            {/** Render the list of items */}
-            {list.map((item) => {
-              return(
-                <div>
-                  {item}
-                </div>
-              );
-            })}
-          </CenterContainer>
-        ) : (
-          <CenterContainer>
-            <h2>No List Items Found</h2>
-          </CenterContainer>
-        )
-      }
-      </div>
+  // fetch list
+  // useEffect(async () => {
+  //   fetch('http://localhost:8000/api/getList')
+  //   .then((res) => res.json())
+  //   .then(list => setList({list}));
+  // })
+  useEffect(async () => {
+    const result = await axios(
+      'http://localhost:8000/api/getList'
     );
-  }
+    setList(result.data);
+  })
+
+  return(
+    <div className="App">
+      <Title>List of Items</Title>
+      {/** Check to see if any items are found */}
+      {list.listArr.length ? (
+        <CenterContainer>
+          {/** Render the list of items */}
+          {list.listArr.map((item) => {
+              <div>
+                {item}
+              </div>
+          })}
+        </CenterContainer>
+      ) : (
+        <CenterContainer>
+          <h2>No List Items Found</h2>
+        </CenterContainer>
+      )
+    }
+    </div>
+  );
 }
 
 export default List;
