+++ client/src/index.js
@@ -1,3 +1,8 @@
+// reference for first 2 imports 
+// https://stackoverflow.com/questions/53558916/babel-7-referenceerror-regeneratorruntime-is-not-defined
+import 'core-js/stable';
+import 'regenerator-runtime/runtime';
+
 import React from 'react';
 import { render } from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
